% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_svelte.R
\name{add_app}
\alias{add_app}
\title{Add app to list of unique apps}
\usage{
add_app(name)
}
\arguments{
\item{name}{:: String, name of the app to check for uniqueness in the
current session. New names will be added to a vector in an
environemt called \code{svelte_apps}.}
}
\value{
Boolean value for if the app is a new app.
}
\description{
Add app to list of unique apps
}
