% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedd_svelte.R
\name{include_svelte}
\alias{include_svelte}
\title{Embed a Svelte App}
\usage{
include_svelte(path, rebuild = FALSE, props)
}
\arguments{
\item{path}{:: String, path to the svelte app main directory}
}
\value{
Code that is inserted into the output document
}
\description{
Use this function to embed a svelte app from a directory / file-path into your
output document (needs to be html).
}
