% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_most_recent_file}
\alias{load_most_recent_file}
\title{Load the Most Recent File of a Specified Type from a Folder}
\usage{
load_most_recent_file(folder, filetype, pattern = NULL)
}
\arguments{
\item{folder}{A character string specifying the path to the folder containing the files.}

\item{filetype}{A character string indicating the file type to search for (e.g., "csv", "sav").}

\item{pattern}{An optional character string representing a regular expression to filter the files by name.
If \code{NULL}, no filtering based on filename will occur.}
}
\value{
If a supported file is found, the function returns the loaded data as a data frame
(or a tibble if reading a CSV file). If no supported file is found or if the file type
is unsupported, the function will return \code{NULL} or the name of the most recent file.
}
\description{
This function searches a specified folder for files of a given type (e.g., CSV, SAV)
and loads the most recently modified file. Optionally, the function can also
filter files based on a specified pattern in their filenames. If no files
are found, or if the file type is unsupported, the function will return an appropriate message.
}
\details{
The function currently supports loading SPSS (.sav) and CSV (.csv) files using the
\code{haven::read_spss()} and \code{readr::read_csv()} functions, respectively. If an unsupported
file type is specified, the function returns the name of the most recent file instead of
loading the data.
}
\examples{
\dontrun{
# Load the most recent CSV file in the "data" folder
data <- load_most_recent_file("data", "csv")

# Load the most recent SAV file in the "data" folder that contains "experiment" in its filename
data <- load_most_recent_file("data", "sav", pattern = "experiment")
}
}
