% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{make_piechart}
\alias{make_piechart}
\title{Create a Pie Chart Based on a Continuous Variable Categorized into Levels}
\usage{
make_piechart(
  df,
  DV_input,
  my_title,
  bins = c(0, 3.501, 4.501, 6.5),
  mylevels = c("Disagree", "Slightly Agree", "Agree")
)
}
\arguments{
\item{df}{A dataframe containing the dependent variable.}

\item{DV_input}{A string representing the name of the dependent variable column in the dataframe.}

\item{my_title}{A string for the title of the pie chart.}

\item{bins}{A numeric vector of breakpoints used to categorize the dependent variable into levels. Default is c(0, 3.6, 4.6, 6.5).}

\item{mylevels}{A character vector representing the names of the levels for the categorical dependent variable. Default is c("Disagree", "Slightly Agree", "Agree").}
}
\value{
A ggplot object displaying a pie chart with the specified title.
}
\description{
This function creates a pie chart from a dataframe by categorizing a continuous dependent variable into
different levels (e.g., Disagree, Slightly Agree, Agree) using custom bins, then visualizes the distribution
of the data in those levels.
}
\examples{
df <- data.frame(score = c(2.5, 4.0, 5.5, NA, 3.9))
make_piechart(df, "score", "Sample Pie Chart")
}
