% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics.R
\name{process_qualtrics_loop_data}
\alias{process_qualtrics_loop_data}
\title{Process Qualtrics Loop Data}
\usage{
process_qualtrics_loop_data(
  dataframe,
  number_range = NULL,
  questions_to_remove = NULL,
  loop_cols = c("loop_group", "loop_item")
)
}
\arguments{
\item{dataframe}{The original dataframe to process.}

\item{number_range}{An optional vector of numbers (after "A") indicating which columns to process. Defaults to all numbers.}

\item{questions_to_remove}{An optional character vector of questions to remove.}

\item{loop_cols}{An optional character vector of length 2, specifying the names of the resulting columns describing the loop item. Defaults to \code{c("loop_group", "loop_item")}. If these names already exist in the dataframe, suffixes \code{"_N"} will be added to make them unique.}
}
\value{
A reformatted long dataframe.
}
\description{
This function processes SPSS data downloaded from Qualtrics when questions are based on a loop.
It separates out the loop row (prefix) and the field in each loop (suffix), reformats the data,
and outputs a long dataframe.
}
\examples{
\dontrun{
processed_data <- process_qualtrics_loop_data(
  dataframe = d_practices_orig,
  questions_to_remove = c("my_loop", "Practice_timing_1", "Practice_timing_2"),
  loop_cols = c("loop_group", "loop_item")
)
}

}
