% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{renderSvelteList}
\alias{renderSvelteList}
\title{Render a svelte list.}
\usage{
renderSvelteList(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{Expression to evaluate. Needs to return a \verb{[svelteList()]}.}

\item{env}{The parent environment for the expression.}

\item{qutoed}{Is the expression quoted?}
}
\description{
Render a svelte list.
}
