% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{svelteButton}
\alias{svelteButton}
\title{Svelte Button}
\usage{
svelteButton(id, label)
}
\arguments{
\item{id}{The id for button.}

\item{label}{The label for the button}
}
\description{
An example input using a svelte component.
}
