% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{svelteList}
\alias{svelteList}
\title{Svelte List}
\usage{
svelteList(items)
}
\arguments{
\item{items}{The items to include in the list.}

\item{id}{The id for list.}
}
\description{
An example output using a svelte component.
}
