% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{write_dev_props}
\alias{write_dev_props}
\title{Write Props for Development}
\usage{
write_dev_props(path, props)
}
\arguments{
\item{path}{:: String, File path to the svelte-app publish directory. e.g. "svelte-app/docs"
or "svelte-app/public"}

\item{props}{:: Any, Any R object convertible into JSON to pass to the app}
}
\value{

}
\description{
Write a .js file to be read in by index.html.
It defines a JS object called props that can be filled
with any R object convertible into JSON.
}
